+++ src/components/Footer.js
@@ -8,10 +8,16 @@ export default function Footer() {
   return (
     <div className="footer-wrapper">
       <div className="socials">
-        <a href="https://github.com/gniarchos/TVTime-ShowsTracker">
+        <a
+          href="https://github.com/gniarchos/TVTime-ShowsTracker"
+          aria-label="github"
+        >
           <Icon icon="mdi:github" className="social-icon" width={35} />
         </a>
-        <a href="https://www.linkedin.com/in/giannis-niarchos/">
+        <a
+          href="https://www.linkedin.com/in/giannis-niarchos/"
+          aria-label="linkedin"
+        >
           <Icon icon="mdi:linkedin" className="social-icon" width={35} />
         </a>
       </div>
@@ -21,16 +27,39 @@ export default function Footer() {
           href="https://gniarchos.github.io/portfolio/"
           target="_blank"
           className="copyright link-portfolio"
+          aria-label="portfolio link"
         >
           Giannis Niarchos
         </a>
-        <h4 className="copyright"> © 2022-2023 </h4>
+        <p className="copyright"> © 2022-2023 </p>
       </div>
 
       <div className="footer-api">
-        <h4>Powered by </h4>
-        <img className="api-images" src={api_logo} alt="api-logo" />
-        <img className="api-images" src={mdblist_logo} alt="api-logo" />
+        <p>Powered by </p>
+        <a
+          href="https://www.themoviedb.org/documentation/api"
+          target="_blank"
+          aria-label="themoviedb api"
+        >
+          <img
+            className="api-images"
+            src={api_logo}
+            alt="api-logo"
+            width="50"
+          />
+        </a>
+        <a
+          href="https://rapidapi.com/linaspurinis/api/mdblist/"
+          target="_blank"
+          aria-label="rapidapi api"
+        >
+          <img
+            className="api-images"
+            src={mdblist_logo}
+            alt="api-logo"
+            width="50"
+          />
+        </a>
       </div>
     </div>
   )